Refactor: introduce useTracking hook and use it across all Pages Router apps
- Create packages/hooks/src/useTracking.ts that handles GA init, global
  listeners, page view tracking on route changes, and exposes trackEvent
- Replace direct analytics function calls with useTracking() in:
  dsayatra, resume-yatra, prep-yatra, oncampus, platform
- Export useTracking from @tbe/hooks package
- Apps using App Router (techyatra, resources, quizes) continue using
  AnalyticsWrapper component which internally uses @tbe/utils

Agent-Logs-Url: https://github.com/The-Boring-Education/TBE-Web/sessions/d384af18-1a12-42f6-82dd-161f373c2671
