+++ apps/dsayatra/src/pages/_app.tsx
@@ -1,15 +1,10 @@
 import "@tbe/components/styles/common.css";
 import "@/index.css";
 
-import {
-  initGA,
-  installGlobalAnalyticsListeners,
-  trackPageview,
-} from "@tbe/components/analytics";
+import { useTracking } from "@tbe/hooks";
 import type { AppProps } from "next/app";
 import Head from "next/head";
 import { useRouter } from "next/router";
-import { useEffect } from "react";
 
 import DsaDashboardLayout from "@/components/layout/DsaDashboardLayout";
 import Layout from "@/components/layout/Layout";
@@ -22,14 +17,7 @@ function AppWithShell({ Component, pageProps }: AppProps) {
   const router = useRouter();
   const useDsaShell = DSA_APP_SHELL_PATHS.has(router.pathname);
 
-  useEffect(() => {
-    initGA();
-    installGlobalAnalyticsListeners();
-
-    const handleRouteChange = (url: string) => trackPageview(url);
-    router.events.on("routeChangeComplete", handleRouteChange);
-    return () => router.events.off("routeChangeComplete", handleRouteChange);
-  }, [router.events]);
+  useTracking();
   const page = <Component {...pageProps} />;
 
   if (useDsaShell) {
