+++ apps/resume-yatra/src/pages/_app.tsx
@@ -3,34 +3,20 @@ import "@/styles/globals.css";
 
 import { AuthProvider } from "@tbe/auth";
 import {
-  initGA,
-  installGlobalAnalyticsListeners,
-  trackPageview,
-} from "@tbe/components/analytics";
-import {
   RadixToaster,
   Toaster as SonnerToaster,
   TooltipProvider,
 } from "@tbe/components/ui";
+import { useTracking } from "@tbe/hooks";
 import { TBEQueryProvider } from "@tbe/query";
 import type { AppProps } from "next/app";
 import Head from "next/head";
-import { useRouter } from "next/router";
-import React, { useEffect } from "react";
+import React from "react";
 
 import { OnboardingCheck } from "@/components/OnboardingCheck";
 
 export default function App({ Component, pageProps }: AppProps) {
-  const router = useRouter();
-
-  useEffect(() => {
-    initGA();
-    installGlobalAnalyticsListeners();
-
-    const handleRouteChange = (url: string) => trackPageview(url);
-    router.events.on("routeChangeComplete", handleRouteChange);
-    return () => router.events.off("routeChangeComplete", handleRouteChange);
-  }, [router.events]);
+  useTracking();
 
   return (
     <>
